Revert "Add API"
