+++ gulpfile.babel.js
@@ -48,7 +48,7 @@ gulp.task('clean', () => del([distDir]));
 
 gulp.task('lint', () =>
   gulp
-  .src(['./*.js', './engine/*.js', './routes/*.js', './src/js/*.js', './tests/**/*.js'])
+  .src(['./*.js', './engine/*.js', './src/js/*.js', './tests/**/*.js'])
   .pipe(plugins.eslint())
   .pipe(plugins.eslint.format())
   .pipe(plugins.eslint.failOnError())
