+++ tests/config/intern-node.js
@@ -4,7 +4,7 @@ define({
     },
   },
 
-  suites: ['tests/nodeUnit', 'tests/redisUnit', 'tests/apiUnit'],
+  suites: ['tests/nodeUnit', 'tests/redisUnit'],
 
   // A regexp to exclude from code coverage calculations
   excludeInstrumentation: /^(?:tests|node_modules)\//,
